Add a prepack script that rewrites README images to be permanent URLs.
This way we don't need to publish images in our NPM tarballs (they're
already in our .npmignore files), but the READMEs will still render with
images on npmjs.com, webcomponents.org, etc.

It resolves master to a SHA, and uses that in a
raw.githubusercontent.com URL, so that if we change or move images files
around, any given version of the README published to NPM won't be
affected.