Fix broken checked property in <mwc-radio>
Previously, we didn't trigger the SelectionController (the class that
synchronizes radio button groups separated by a shadow root) when the
checked property was set, so the previously checked radio in a group was
not unchecked when a new one was checked.

This was more complicated than just triggering it from the existing
property @observer, because that gives us batched async changes (via the
UpdatingElement updated callback), so we can't reconstruct the order of
checked sets correctly. Replaces the observer with a setter that can
handle this synchronously.