Use SelectionController even without native shadow.
Previously, we did not use the SelectionController class when we were
not using native shadow roots, because we could rely on the native
behavior to manage selection groups for us.

However, since the native <input> is not stamped until the async first
update, groups were not being managed correctly immediately after
construction.

Now, we always use a SelectionController. It's probably a little less
efficient when we don't have native shadow, and it would be possible to
e.g. force an early render to make this work, but I don't think the
extra complexity is worth it.