Fix bug where mwc-snackbar labelText could not be set.
This adds a lit directive implementation of
@material/mdc-snackbar/util.ts#announce, which does some tricks to
ensure that snackbar labels will be handled correctly by screen readers.

The existing MDC announce util function is difficult to use directly
here, because Lit can crash when DOM that it is managing changes outside
of its purvue. In this case, we would render our labelText as the text
content of the label div, but the MDC announce function then clears that
text content, and resets it after a timeout. We do the same thing here,
but in a way that fits into Lit's lifecycle.

Fixes https://github.com/material-components/material-components-web-components/issues/367