Prevent mwc-list-item selectedText from including text from font ligatures of slotted icons
Problem: the `selectedText` property of mwc-list-item returns the text content of all slots, including the font ligatures of slotted mwc-icons. This currently prevents us from using slotted icons with mwc-list-items inside of an mwc-select because the displayed `selectedText` will also include the font ligatures' text.

Fix: Return only the text of the default/primary slot, not that of the icon slots.

I also added "Options with Icons" to the mwc-select demo page.