Set `aria-labelledby` conditionally based on the presence of an assigned `label`
This change fixes a dangling reference issue in the `aria-labelledby` attribute set in various MWC element internals by only populating the `aria-labelledby` attribute when a `label` value has actually been provided to the outer MWC element itself.

Previously, the `aria-labelledby` attribute is unconditionally set to the id `label`, which is expected to point to the element created in the shadow dom which will be populated with the label provided to the outer MWC element. However, that `label` element is only created when a `label` property is actually provided, so the `aria-labelledby` attribute would end up mistakenly pointing to a non-existent element.

PiperOrigin-RevId: 381519765