Remove definition of invalid `aria-errortext` property
The `aria-errortext` attribute currently populated in the <mwc-textfield> isn't actually a valid attribute according to the spec.

It's possible that the original intent was to assign an `aria-errormessage` (https://www.w3.org/TR/wai-aria/#aria-errormessage), but given that the single `helper-text` element is used to surface descriptive text *and* validation messages, defining an `aria-errormessage` attribute can result in duplicate announcement of the helper text by screen readers (once for the describedby, and again for errormessage), which doesn't seem ideal.

For this reason, this change opts to just remove the invalid `aria-errortext` property instead of replacing it with `aria-errormessage`.

PiperOrigin-RevId: 382617574