(chore) Fix Typescript 4.4 compilation issues with aria properties
Typescript 4.4 defines aria properties on Element prototype to be non-nullable. This matches the spec language in https://www.w3.org/TR/wai-aria-1.2/#AriaAttributes, but Chrome and Safari as of when these properties were added to MWC allowed the property to be set with `null` to clear the attribute.

There should be some ongoing work to align the spec and browsers in behavior, but that is out of scope for this project, unblocking Typescript 4.4 is of higher priority.

Related microsoft/TypeScript-DOM-lib-generator#1119
Fixes #2676

PiperOrigin-RevId: 395581595