`AccessibleSnackbarLabel`: Avoid using `render` on an element that must have its `textContent` repeatedly set.
Setting `.textContent` of a node that was previously rendered to with Lit's `render` function will remove the comments Lit uses to keep track of the region it controls inside that node. In an upcoming version of Lit, if it detects this situation when dev-mode is enabled (as it is in the tests here), it will throw an error.

To work around this, I changed this directive to always update the content of `labelEl` manually, rather than using `render`. Originally, I planned to switch everything over to `render`, but the comments here imply that setting `.textContent` specifically is a critical part of making screenreaders behave as expected.

PiperOrigin-RevId: 398323705