fix(tab): Fixes how tabs renders icons.
When a image icon is provided through the icon `<slot>`, it is now rendered directly into the tab content element (technically into an icon wrapper that is a direct child of the tab content element). Previously the slotted icon would be rendered as the contents of `<md-icon>`. While this didn't affect the visual output of the tab component, it created unnecessary additional layers of nesting/shadow dom.

PiperOrigin-RevId: 466052019