fix(menu): properly implement selected state
Before, I set the focused, actively-keyboard-interacted, or pressed element as visual selected state. That doesn't seem to be how GMDC handles it and created problems down the line with md-select where we could not show the currently-selected item + keyboard navigation at the same time.

This CL brings the behavior in line with GMDC and fixes this problem down the line for select.

Also in this CL I add the selected state to sub-menu-item when it's open

PiperOrigin-RevId: 519166067