fix(chips): linear navigation not working in VoiceOver
A lot of the changes are in order to fix filter chips with VoiceOver. With the extra divs and no aria-hidden on ripple/focus ring, VoiceOver got stuck on the first filter chip with linear navigation.

Elevated link chips also got their border changed to white instead of yellow in HCM to match Wiz.

PiperOrigin-RevId: 559908582