chore(behaviors): update validity anchor on first render
This is needed specifically for text fields. There appears to be an odd Chrome/Safari issue where an element that `delegatesFocus: true` to an inner `<input>` MUST provide that input as the validity anchor. Otherwise, when a form submits or calls `form.reportValidity()`, an error will be thrown.

See https://lit.dev/playground/#gist=6c26e418e0010f7a5aac15005cde8bde for a reproduction.

Unfortunately I couldn't programmatically trigger this behavior in a jasmine test.

PiperOrigin-RevId: 586723140