fix(slider): do not use :host-context to determine document position
Because slider uses `host-context([dir="rtl"])` the following code causes slider to get confused:

```
<div dir="rtl">
   <div dir="ltr">
      <md-slider>
```

the html input incorrectly picks up that it's in a ltr context but the track css thinks it's in rtl since the `:host-context` selector picks up the `dir` from the highest element in the subtree.

This is a possible fix where we just rely on :host([dir="rtl"]) to determine document direction instead.

PiperOrigin-RevId: 590757010